requirements.txt
streamlit==1.32.0
pandas==2.0.3
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.1.0
google-api-python-client==2.118.0
groq==0.4.1
python-dotenv==1.0.0
git+https://github.com/serpapi/google-search-results-python.git
requests==2.31.0
